refactor: modularizes messaging logic code
